+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Balanced ternary computer: encode".
 - Tests for "Extended hamming codes".
 - Tests for "Organic compounds".
+- Tests for "Nature of quadrilaterals".
 
 ### Changed
 - Docker Node image from "14.16.1-alpine" to "node:16.14.2-alpine".